pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ahmadalrasheed <alrasheedahmad97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
